requirements.txt
fastapi==0.115.12
openai==1.108.1
pydantic==2.11.3
python-dotenv==1.1.0
uvicorn[standard]==0.34.2
